fix net rate calc review fixes
